requirements.txt
-i https://pypi.org/simple
boto3==1.26.11
botocore==1.29.11 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
pydantic==1.10.2
git+https://github.com/pynamodb/PynamoDB@2ec2f355eef362086662fecd6248befcf09f183c#egg=pynamodb
python-datauri==1.1.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
s3transfer==0.6.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
